+++ run.rb
@@ -1,0 +1,3 @@
+require_relative 'task_1'
+
+work('data_medium.txt')
